requirements.txt
python-dotenv
azure-identity
azure-keyvault-keys